Adding sample collection code
Adjusted get_temp() to poll until NaN is false. Adding math.h
library. Adding functions to calculate mean and standard deviation of an
array of data. Changed initialization of doubles to 0.0. Function
add_sample() will add a sample to the samples collection if it is not
outside our margin of error (currently either less than 10, more than
500, or more than 5 standard deviations away from the mean), otherwise
it repeats the prior sample as filler. In loop(), internal_temp is
calculated as the mean of the samples. In setup(), the samples array is
filled with current temperature data.